+++ test/unit/adapters/test.sequelize.js
@@ -150,13 +150,26 @@ describe('=> Sequelize adapter', () => {
       { title: 'Tyrael' },
     ].map(opts => Post.create(opts)));
 
-    const { rows, count } = await Post.findAll({
+    const { rows, count } = await Post.findAndCountAll({
       where: {
         title: { $like: '%ea%' },
       },
     });
     assert.equal(rows.length, 1);
     assert.equal(count, 1);
+  });
+
+  it('Model.findOne()', async () => {
+    const posts = await Promise.all([
+      { title: 'Leah' },
+      { title: 'Tyrael' },
+    ].map(opts => Post.create(opts)));
+
+    const post = await Post.findOne(posts[1].id);
+    assert.equal(post.title, 'Tyrael');
+
+    const post2 = await Post.findOne({ title: 'Leah' });
+    assert.equal(post2.title, 'Leah');
   });
 
   it('model.restore()', async () => {